requirements.txt
streamlit==1.31.1
Pillow==10.2.0
reportlab==4.1.0
numpy==1.26.3
opencv-python==4.9.0.80 